+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## 1.1.1
 
 * Fix a bug where including a form module would mess up the options has of the validations (under older Rails).
+* `Form#save` now returns the result of the `model.save` invocation.
 
 ## 1.1.0